requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
requests>=2.31
python-dotenv>=1.0
gunicorn>=22.0
